Added UploadingCompletedEvent. Closes #11.

+++ src/main/java/com/gmail/kompotik/gwtupld/client/UploadProgressHandlers.java
@@ -6,5 +6,5 @@ public interface UploadProgressHandlers {
   public void onProgress(String id, String filename, int loaded, int total);
   public void onComplete(String id, String filename, JSONValue response);
   public void onCancel(String id, String filename);
-  public Long onSubmit(String id, String filename);
+  public int onSubmit(String id, String filename);
 }
